+++ cc3d/__init__.py
@@ -53,6 +53,8 @@ cc3d_scripts_path = abspath(join(cc3d_py_dir, config.cc3d_scripts_path_rel))
 cc3d_lib_shared = abspath(join(cc3d_install_prefix, 'bin'))
 cc3d_lib_static = abspath(join(cc3d_install_prefix, 'lib'))
 
+cc3d_cpp_bin_path_pathlib = Path(cc3d_cpp_bin_path)
+
 os.environ['COMPUCELL3D_STEPPABLE_PATH'] = cc3d_steppable_path + os.sep
 os.environ['COMPUCELL3D_PLUGIN_PATH'] = cc3d_plugin_path + os.sep
 
@@ -72,8 +74,11 @@ if sys.platform.startswith('win'):
 
         if cc3d_lib_shared not in path_env_list:
             os.add_dll_directory(cc3d_lib_shared)
-        if cc3d_cpp_bin_path not in path_env_list:
-            os.add_dll_directory(cc3d_cpp_bin_path)
+        if cc3d_cpp_bin_path_pathlib.exists() and cc3d_cpp_bin_path_pathlib.is_dir():
+            # note this directory may not exist if we are using conda package layout.
+            # It only exists if we compile ti in a standalone mode
+            if cc3d_cpp_bin_path not in path_env_list:
+                os.add_dll_directory(cc3d_cpp_bin_path)
 
         os.add_dll_directory(os.environ['COMPUCELL3D_PLUGIN_PATH'])
         os.add_dll_directory(os.environ['COMPUCELL3D_STEPPABLE_PATH'])
@@ -82,8 +87,11 @@ if sys.platform.startswith('win'):
 
         if cc3d_lib_shared not in path_env_list:
             path_env_list.insert(0, cc3d_lib_shared)
-        if cc3d_cpp_bin_path not in path_env_list:
-            path_env_list.insert(0, cc3d_cpp_bin_path)
+        if cc3d_cpp_bin_path_pathlib.exists() and cc3d_cpp_bin_path_pathlib.is_dir():
+            # note this directory may not exist if we are using conda package layout.
+            # It only exists if we compile ti in a standalone mode
+            if cc3d_cpp_bin_path not in path_env_list:
+                path_env_list.insert(0, cc3d_cpp_bin_path)
 
         path_env_list.insert(0, os.environ['COMPUCELL3D_PLUGIN_PATH'])
         path_env_list.insert(0, os.environ['COMPUCELL3D_STEPPABLE_PATH'])